Fixed forms and comments

+++ app/view.py
@@ -18,15 +18,20 @@ def login():
     form = LoginForm()
     
     if request.method == 'POST' and form.validate_on_submit():
-        user = db.session.query(User).filter(User.name == form.name.data and User.email == form.email.data).first()
-        user_email = db.session.query(User).filter(user.email == form.email.data).first()
+        try:
 
-        if user and user_email :
-            login_user(user, remember=form.remember.data)
-            return redirect('/comments/')
+            user = db.session.query(User).filter(User.name == form.name.data).first()
+            user_email = db.session.query(User).filter(user.email == form.email.data).first()
 
-        flash("Invalid name/email.")
-        return render_template('login.html', form=form)
+            if user and user_email :
+                login_user(user)
+                return redirect('/comments/')
+
+            flash("Invalid name/email.")
+            return render_template('login.html', form=form)
+        except:
+            flash("Invalid name/email.")
+            return render_template('login.html', form=form)
     else:
         
         return render_template('login.html', form=form)
@@ -55,7 +60,6 @@ def add_user():
         if new_birthday == datetime.now():
             new_birthday = None
 
-        new_remember = form.remember.data
         new_user = User(name=new_name, surname=new_surname, email=new_email, gender=new_gender, birthday=new_birthday)
         
         try:
@@ -73,23 +77,25 @@ def add_user():
 @login_required
 def feedback():
     form = AddFeedback()
-    if request.method == 'POST' and form.validate_on_submit():
-        user = db.session.query(User).filter(User.email == form.email.data and User.name == form.email.name ).first()
-        user_name = db.session.query(User).filter(user.name == form.name.data).first()
-        if user and user_name:
-            user_id = user.id
-            content = form.feedback.data
-            new_comment = Comment(user_id=user_id,content=content)
-            try:
-                db.session.add(new_comment)
-                db.session.commit()
-                return redirect('/comments/')
-            except:
-                flash("Eroor with adding new comment")
-                return render_template('addfeedback.html', form=form)
-           
-        flash("Invalid name/email")
-        return render_template('addfeedback.html', form=form)
+    if request.method == 'POST' and form.validate_on_submit(): 
+        try:
+            user = db.session.query(User).filter(User.email == form.email.data).first()
+            user_name = db.session.query(User).filter(user.name == form.name.data).first()
+            if user and user_name:
+                user_id = user.id
+                content = form.feedback.data
+                new_comment = Comment(user_id=user_id,content=content)
+                try:
+                    db.session.add(new_comment)
+                    db.session.commit()
+                    return redirect('/comments/')
+                except:
+                    flash("Eroor with adding new comment")
+                    return render_template('addfeedback.html', form=form)
+        except:
+            flash("Invalid name/email")
+            return render_template('addfeedback.html', form=form)
+    flash("Eroor with adding new comment")
     return render_template('addfeedback.html', form=form)
 
 
@@ -103,11 +109,17 @@ def index():
 @app.route('/delete/<int:id>')
 def delete(id):
     task_to_delete = Comment.query.get_or_404(id)
-    try:
-        db.session.delete(task_to_delete)
-        db.session.commit()
+    if db.session.query(User).filter(current_user.id == task_to_delete.user_id).first():
+        try:
+            db.session.delete(task_to_delete)
+            db.session.commit()
+            return redirect('/comments/')
+        except:
+            return flash("Delete error")
+    else:
+        flash("Delete error")
+
         return redirect('/comments/')
-    except:
-        return 'Delete error'
+
